Fix the issue with relativenumber
It used to be that if I was on Insert mode and lost focus
I would lose the absolute number, which caused me to have the
absolute number after gaining focus back. Leading to undesirable
behavior.

This fixes it by preventing to go the relative number when focus is lost.
Hence keeping things in order.