Register the extra files last
This change allows for overriding configurations with custom ones
on different systems.
An example: an alias may fail to work properly due a path on the current
system being different. Instead of having local changes to the current .alias
file (that will be never committed) we can put those differences in the .extra file